Nation Silhouette
Modeled a simple "NationSilhouette" (May add 2-3 more and have them randomly chosen).
Instead of importing an .fbx, I saved the blend file directly into the Assets folder of the Unity project, so that things can be updated more smoothly.
Parented silhouette prefabs to each nation, and impart the nation's color to it at runtime.
(Tried to add code to scale nations down as they take damage, but it wasn't really working with all the children they have).
Tagged WMD prefab with WMD tag, as forgetting to do so was preventing them from actually doing damage to the nations.