Nation Revival
Instead of nations being destroyed when they run out of HP, they now assume the color of the nation who "struck the final blow to conquer them".
When they are "Revived", they regain full health/scale and continue playing.
When all nations share the same color (on their bases, that is. I still haven't gotten the silhouettes to change color), the game is over.